Specify version of log
Even though most of the other dependencies are for in-dev crates/dev branches,
there is no need to be unspecific about the version of the log.

Also, I despise wildcard dependencies and consider them one of the few solid
pieces of evidence that Satan exits (and hates SemVer). Just my opinion though.