Add streaming body support to proto::pipeline
The two major components of this commit are:

* Extracting common logic from `pipeline::Client` and `pipeline::Server` into
  a reusable component.
* Support streaming bodies for both requests and responses.

The assumption for pipeline based protocols is that a request or response may
be followed by an arbitrary number of body frames. In order to avoid waiting
for the entire body to arrive and having to buffer the contents in memory, the
body is able to be streamed to the service by including a `Stream<T>` with the
request / response. T is an arbitrary type that represents a single body
frame.