Replace the `reactor` module with `tokio-core`
This commit replaces tokio's `reactor` module with the `tokio-core` crate,
obviating the need for the `util` module of tokio as well (primitives and
functionality now added to the `futures` crate). Discussed with Carl, this
should help not only simplify tokio but also help plumb futures/streams up and
down through tokio as well!

The recently-created `tokio-core` is an extraction of the `futures-mio` crate
from the `futures-rs` repository, and `tokio-core` is now the "source of truth"!