Change how outbound body streams are handled
The transport is no longer responsible for creating a channel for the
streaming body. Instead, a flag is set in the frame and the dispatcher creates
the channel.

This means that the `Frame` enum in both pipeline and multiplex has changed.
It also means that the outbound message being dispatched must be of
`Message<T>` type vs. only `T`.

This commit also includes an internal overhaul of the multiplexer.