Add an `easy` module for no streaming support
This commit adds a new module to `tokio-proto`, `easy`, which like
`tokio_core::easy` is intended to help you get off the ground running quickly
for any particular protocol.

This module mirrors the API of the pipeline/multiplex modules with a
server/client implementation that's built on top of the actual server/client in
the top-level modules. The types and generics, however, are greatly reduced to
help work with them ergonomically.