Generalize TcpServer::serve's service a bit
The services accepted by `TcpServer::serve` wire up the service's
request/response types the underlying request/response types of the `BindServer`
implementation. These types tend to be what we want for the simple case (non
streaming), but for the streaming case they have the
`tokio_proto::streaming::Message` type wired in.

The `Message` type typically isn't what libraries want as the request/response
types for their services, so this PR generalizes to use `From` and `Into` to
ensure that custom types can still be used so long as they can be converted.

I think this sould be backwards compatible as well (not causing inference
regressions), hopefully at least.