Use `E: Executor` instead of `tokio_core`'s `Handle`
This commit moves the `BindClient` and `BindServer` traits to working with the
new `Executor` trait in the `futures` crate. The intention here is to abstract
over the ability to spawn a task and remove the hardwired dependency on
`tokio_core`, allowing `tokio-proto` to be used with other event loops and/or
reactors.

Technically this is a breaking change but I think in practice this won't be (and
hence we may be able to land in the 0.1 series). No other external
implementations of `Bind{Server,Client}` are expected so the signature change
should be find and the only previously provided type, `Handle`, should implement
`Executor` to continue working.